+++ .storybook/chsk.ts
@@ -26,9 +26,8 @@ const removeUnnecessaryProps = (input: string) => {
     unnecessaryProps.forEach(prop => {
         result = result
             .replace(prop + '={() => {}}', '')
-            .replace(prop + '={function noRefCheck() {}}', '')
+            .replace(prop + '={function noRefCheck(){}}', '')
     })
-    //onClick={function noRefCheck() {}}
     unnecessaryStrings.forEach(target => {
         result = result.replace(target, '')
     })
@@ -52,9 +51,7 @@ const prettierChartCode = (input: string) => {
 
 /** custom code transforms specific to the chsk project */
 export const transformCode = (input: string) => {
-    console.log('input: ' + input)
     const shortInput = removeUnnecessaryProps(input)
-    console.log('short: ' + shortInput)
     try {
         return prettierCode(shortInput)
     } catch {}